backend/requirements.txt
# Core
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
pydantic>=2.9.0
pydantic-settings>=2.5.0

# Banco de dados
sqlalchemy[asyncio]>=2.0.35
asyncpg>=0.30.0
alembic>=1.14.0

# Autenticação
python-jose[cryptography]>=3.3.0
bcrypt>=4.2.0

# Criptografia de conexões
cryptography>=43.0.0

# HTTP client (Power BI)
httpx>=0.27.0

# Validação de email
email-validator>=2.2.0

# Testes
pytest>=8.3.0
pytest-asyncio>=0.24.0
httpx>=0.27.0

# Opcional: Redis
# redis>=5.0.0
# fastapi-cache2>=0.2.0
